Update 2015-10-27-vim-and-python-a-match-made-in-heaven.markdown
Addressing the fact that the output has +python/dyn and -python3 and doesn't actually have a plain +python which the learner was directed to look for in the output. Also, if I recompile vim with to have +python3 would that be okay or should it be avoided for some reason?